+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ---
 
 - Configure Faraday's default adapter as `net_http`
+- Fix `X-Client-Start` & `X-Client-Timeout` headers to be strings not integers
 
 2.17
 ---